renamed dashboard folder
